Refactor ProjectCreateModal.react
Refactoring the ProjectCreateModal to use a reusable component ProjectCreateView. This component will be reused in the InstanceLaunchModal when a project doesn't exist.
